+++ src/main/java/me/pajic/tiered_backpacks/platform/Platform.java
@@ -2,7 +2,6 @@ package me.pajic.tiered_backpacks.platform;
 
 import me.pajic.tiered_backpacks.ui.BackpackMenu;
 import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
-import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.MenuType;
 import net.minecraft.world.item.ItemStack;
@@ -21,5 +20,5 @@ public interface Platform {
 
 	MenuType<BackpackMenu> constructMenu();
 
-	InteractionResult openBackpackScreen(Player player, ItemStack backpack);
+	void openBackpackScreen(Player player, ItemStack backpack);
 }