+++ src/main/java/me/pajic/tiered_backpacks/util/BackpackClientUtil.java
@@ -6,6 +6,7 @@ import me.pajic.tiered_backpacks.compat.TrinketsCompat;
 import me.pajic.tiered_backpacks.keybind.ModKeybinds;
 import me.pajic.tiered_backpacks.network.ModNetworking;
 import net.minecraft.client.Minecraft;
+import net.minecraft.sounds.SoundEvents;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.item.ItemStack;
 
@@ -16,13 +17,18 @@ public class BackpackClientUtil {
 			ItemStack chestItem = ItemStack.EMPTY;
 			if (CompatFlags.TRINKETS_LOADED) chestItem = TrinketsCompat.tryGetTrinketBackpack(client.player);
 			if (CompatFlags.OHMEGA_LOADED) chestItem = OhmegaCompat.tryGetOhmegaBackpack(client.player);
-			if (!chestItem.isEmpty()) TieredBackpacks.xplat().sendToServer(new ModNetworking.C2SOpenBackpackPayload(2));
+			if (!chestItem.isEmpty()) {
+				client.player.playSound(SoundEvents.BUNDLE_INSERT);
+				TieredBackpacks.xplat().sendToServer(new ModNetworking.C2SOpenBackpackPayload(2));
+			}
 			else {
 				chestItem = client.player.getItemBySlot(EquipmentSlot.CHEST);
 				if (BackpackUtil.isValidContainerHolder(chestItem)) {
+					client.player.playSound(SoundEvents.BUNDLE_INSERT);
 					TieredBackpacks.xplat().sendToServer(new ModNetworking.C2SOpenBackpackPayload(1));
 				} else if (TieredBackpacks.CONFIG.canOpenFromInventory.get()) {
 					if (client.player.getInventory().getNonEquipmentItems().stream().anyMatch(stack -> stack.is(BackpackUtil.BACKPACKS))) {
+						client.player.playSound(SoundEvents.BUNDLE_INSERT);
 						TieredBackpacks.xplat().sendToServer(new ModNetworking.C2SOpenBackpackPayload(0));
 					}
 				}